// src/types/enums.ts
export enum SelectionMode {
  None = 'none',
  Rectangular = 'rectSelection',
  Move = 'move',
  Scale = 'scale',
  Direct = 'directSelection',
}

export enum HandlePosition {
  TopLeft = 'topLeft',
  TopCenter = 'topCenter',
  TopRight = 'topRight',
  BottomRight = 'bottomRight',
  BottomCenter = 'bottomCenter',
  BottomLeft = 'bottomLeft',
  LeftCenter = 'leftCenter',
  RightCenter = 'rightCenter',
}